REFACTOR: Remove bitdubai from wallet assets user wallet plugin. #5808
