+++ DAP/plugin/wallet/fermat-dap-plugin-wallet-assets-user-wallet/src/main/java/org/fermat/fermat_dap_plugin/layer/wallet/asset/user/developer/bitdubai/version_1/structure/exceptions/CantExecuteAssetUserTransactionException.java
@@ -1,0 +1,20 @@
+package org.fermat.fermat_dap_plugin.layer.wallet.asset.user.developer.bitdubai.version_1.structure.exceptions;
+
+import com.bitdubai.fermat_api.FermatException;
+
+/**
+ * Created by franklin on 05/10/15.
+ */
+public class CantExecuteAssetUserTransactionException extends FermatException {
+    /**
+     * This is the constructor that every inherited FermatException must implement
+     *
+     * @param message        the short description of the why this exception happened, there is a public static constant called DEFAULT_MESSAGE that can be used here
+     * @param cause          the exception that triggered the throwing of the current exception, if there are no other exceptions to be declared here, the cause should be null
+     * @param context        a String that provides the values of the variables that could have affected the exception
+     * @param possibleReason an explicative reason of why we believe this exception was most likely thrown
+     */
+    public CantExecuteAssetUserTransactionException(String message, Exception cause, String context, String possibleReason) {
+        super(message, cause, context, possibleReason);
+    }
+}
